Add new dashboard widgets: Gauge, Heatmap, Histogram, Line Chart, Pie Chart, Stat, Table
- Implemented GaugeWidget for displaying gauge metrics with thresholds.
- Created HeatmapWidget to visualize data density across two dimensions.
- Developed HistogramWidget for displaying frequency distribution of data.
- Added LineChartWidget for rendering line charts with brush functionality.
- Introduced PieChartWidget for visualizing proportions of a whole.
- Built StatWidget to show a single statistic with optional trend indicators.
- Implemented TableWidget for displaying tabular data with sorting and pagination.
- Added widgetConstants for shared constants and utility functions across widgets.
